+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/assets/registry.py
@@ -274,55 +274,3 @@ def oss_destinations_dataframe(oss_registry_from_metadata: ConnectorRegistryV0) 
     destinations = oss_registry_from_metadata_dict["destinations"]
     return output_dataframe(pd.DataFrame(destinations))
 
-
-# Old Registry
-
-
-@asset(group_name=GROUP_NAME)
-def legacy_cloud_sources_dataframe(legacy_cloud_registry_dict: dict) -> OutputDataFrame:
-    sources = legacy_cloud_registry_dict["sources"]
-    return output_dataframe(pd.DataFrame(sources))
-
-
-@asset(group_name=GROUP_NAME)
-def legacy_oss_sources_dataframe(legacy_oss_registry_dict: dict) -> OutputDataFrame:
-    sources = legacy_oss_registry_dict["sources"]
-    return output_dataframe(pd.DataFrame(sources))
-
-
-@asset(group_name=GROUP_NAME)
-def legacy_cloud_destinations_dataframe(legacy_cloud_registry_dict: dict) -> OutputDataFrame:
-    destinations = legacy_cloud_registry_dict["destinations"]
-    return output_dataframe(pd.DataFrame(destinations))
-
-
-@asset(group_name=GROUP_NAME)
-def legacy_oss_destinations_dataframe(legacy_oss_registry_dict: dict) -> OutputDataFrame:
-    destinations = legacy_oss_registry_dict["destinations"]
-    return output_dataframe(pd.DataFrame(destinations))
-
-
-@asset(required_resource_keys={"legacy_cloud_registry_gcs_blob"}, group_name=GROUP_NAME)
-def legacy_cloud_registry(legacy_cloud_registry_dict: dict) -> ConnectorRegistryV0:
-    return ConnectorRegistryV0.parse_obj(legacy_cloud_registry_dict)
-
-
-@asset(required_resource_keys={"legacy_oss_registry_gcs_blob"}, group_name=GROUP_NAME)
-def legacy_oss_registry(legacy_oss_registry_dict: dict) -> ConnectorRegistryV0:
-    return ConnectorRegistryV0.parse_obj(legacy_oss_registry_dict)
-
-
-@asset(required_resource_keys={"legacy_cloud_registry_gcs_blob"}, group_name=GROUP_NAME)
-def legacy_cloud_registry_dict(context: OpExecutionContext) -> dict:
-    oss_registry_file = context.resources.legacy_cloud_registry_gcs_blob
-    json_string = oss_registry_file.download_as_string().decode("utf-8")
-    oss_registry_dict = json.loads(json_string)
-    return oss_registry_dict
-
-
-@asset(required_resource_keys={"legacy_oss_registry_gcs_blob"}, group_name=GROUP_NAME)
-def legacy_oss_registry_dict(context: OpExecutionContext) -> dict:
-    oss_registry_file = context.resources.legacy_oss_registry_gcs_blob
-    json_string = oss_registry_file.download_as_string().decode("utf-8")
-    oss_registry_dict = json.loads(json_string)
-    return oss_registry_dict
