+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/templates/connector_registry_locations.html
@@ -3,7 +3,7 @@
 {% block title%}Airbyte Connector Registry{% endblock %}
 
 {% block content %}
-<h1>Connector Registries</h1>
+<h1>Airbyte Connector Registries</h1>
 <h2>Sources</h2>
 {{ sources_table }}
